Efficient threading of pangloss
Threads are only spawned once, with the optimal number of jobs.
They access the same matrix, but this should be the only bottleneck

+++ src/panCommon.cpp
@@ -169,7 +169,7 @@ arma::vec dlib_to_arma(const column_vector& dlib_vec)
 {
    arma::vec converted(dlib_vec.nr());
 
-   for (size_t i = 0; i < dlib_vec.nr(); ++i)
+   for (unsigned int i = 0; i < dlib_vec.nr(); ++i)
    {
       converted(i) = dlib_vec(i);
    }
@@ -182,7 +182,7 @@ column_vector arma_to_dlib(const arma::vec& arma_vec)
    column_vector converted;
    converted.set_size(arma_vec.n_elem);
 
-   for (size_t i = 0; i < arma_vec.n_elem; ++i)
+   for (unsigned int i = 0; i < arma_vec.n_elem; ++i)
    {
       converted(i) = arma_vec(i);
    }
